kim_flask/requirements.txt
alembic==1.4.2
attrs==19.3.0
autopep8==1.5.2
blinker==1.4
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.16
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
